+++ crates/execution/src/python/mod.rs
@@ -15,6 +15,7 @@
 
 //! Python bindings from [PyO3](https://pyo3.rs).
 
+pub mod config;
 pub mod fee;
 pub mod fill;
 pub mod latency;
@@ -45,6 +46,8 @@ pub fn execution(_: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
         reconciliation::py_create_position_reconciliation_venue_order_id,
         m
     )?)?;
+    m.add_class::<crate::engine::config::ExecutionEngineConfig>()?;
+    m.add_class::<crate::order_emulator::config::OrderEmulatorConfig>()?;
     m.add_class::<crate::models::fee::FixedFeeModel>()?;
     m.add_class::<crate::models::fee::MakerTakerFeeModel>()?;
     m.add_class::<crate::models::fee::PerContractFeeModel>()?;